feat(pumbility): widen the stored columns to float
A PUMBILITY pool is fifty per-chart values that each carry a real
fraction, so an integer column spends precision before anything has
decided to. Seven columns widen: the four PlayerStats pools, the two
PlayerHistory ones, and the per-score gain on ScoreHighlight.

Up is a widening and runs in place -- no rewrite, no backfill, and
existing rows read back with a zero fraction until the next ratings
sweep. The records still carry Rating/int, so the repositories cast on
read and nothing on screen moves yet.

+++ ScoreTracker/ScoreTracker.PlayerProgress/Infrastructure/EFPlayerStatsRepository.cs
@@ -121,10 +121,11 @@ namespace ScoreTracker.PlayerProgress.Infrastructure
                     return new PlayerStatsRecord(userId, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1);
 
                 return new PlayerStatsRecord(entity.UserId, entity.TotalRating, entity.HighestLevel, entity.ClearCount,
-                    entity.CoOpRating,
-                    entity.AverageCoOpScore, entity.SkillRating, entity.AverageSkillScore, entity.AverageSkillLevel,
-                    entity.SinglesRating,
-                    entity.AverageSinglesScore, entity.AverageSinglesLevel, entity.DoublesRating,
+                    (int)entity.CoOpRating,
+                    entity.AverageCoOpScore, (int)entity.SkillRating, entity.AverageSkillScore,
+                    entity.AverageSkillLevel,
+                    (int)entity.SinglesRating,
+                    entity.AverageSinglesScore, entity.AverageSinglesLevel, (int)entity.DoublesRating,
                     entity.AverageDoublesScore, entity.AverageDoublesLevel, entity.CompetitiveLevel,
                     entity.SinglesCompetitiveLevel, entity.DoublesCompetitiveLevel,
                     entity.EstimatedPumbilityRank, entity.EstimatedSinglesPumbilityRank,
@@ -140,10 +141,11 @@ namespace ScoreTracker.PlayerProgress.Infrastructure
             return await database.Set<PlayerStatsEntity>()
                 .Where(s => userIds.Contains(s.UserId) && s.MixId == mixId).Select(entity =>
                     new PlayerStatsRecord(entity.UserId, entity.TotalRating, entity.HighestLevel, entity.ClearCount,
-                        entity.CoOpRating,
-                        entity.AverageCoOpScore, entity.SkillRating, entity.AverageSkillScore, entity.AverageSkillLevel,
-                        entity.SinglesRating,
-                        entity.AverageSinglesScore, entity.AverageSinglesLevel, entity.DoublesRating,
+                        (int)entity.CoOpRating,
+                        entity.AverageCoOpScore, (int)entity.SkillRating, entity.AverageSkillScore,
+                        entity.AverageSkillLevel,
+                        (int)entity.SinglesRating,
+                        entity.AverageSinglesScore, entity.AverageSinglesLevel, (int)entity.DoublesRating,
                         entity.AverageDoublesScore, entity.AverageDoublesLevel, entity.CompetitiveLevel,
                         entity.SinglesCompetitiveLevel, entity.DoublesCompetitiveLevel,
                         entity.EstimatedPumbilityRank, entity.EstimatedSinglesPumbilityRank,
